backend/requirements.txt
fastapi[standard]==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.31
aiosqlite==0.20.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
google-auth==2.29.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
email-validator==2.1.1
